+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for error-codes
 
+## 0.1.2.0 -- 2023-09-07
+
+* Add `enumString`, `enumByteArray`, `enumShortText`, `enumText`.
+* Add `descriptionString`, `descriptionByteArray`, `descriptionShortText`, `descriptionText`.
+
 ## 0.1.1.0 -- 2023-07-13
 
 * Add `Foreign.C.Error.Pattern`.